Add mlpack
This is currently an 'MVP' -- a minimally viable product (demoo) wrapping two
algorithms.  The mlpack build is itself reliant on cmake, and we do a
two-line patch to one dependency (cereal) there.  This currently fails here
leading to the two macOS build issues which "should be" fixable.  Webassembly
should work as it does for the CRAN package of mlpack at r-universe -- we
will look into this too. mingw is probably out of scope but standard windows
may work. To be seen.