Support opt-in platforms
This PR adds support for `opt_in_platforms` attribute to the extension's
`description.yml` file. This attribute is optional, it can contain the
semicolon-separated list of platforms that have `"opt_in": true` set in
the `distribution_matrix.json`, for example: 'windows_arm64;'. It is
handled the same way as existing `excluded_platforms` attribute.

Additionally `more_opt_in` input parameter is added to `build.yml`
workflow, it is handled the same way as the existing `more_excluded`
parameter and is intended to be used to test new opt-in architectures.

Fixes: #935